split machine.rs into compiler,machine and program
