+++ src/tests.rs
@@ -1,11 +1,9 @@
-use regex_syntax::{hir::visit, Parser};
-
-use crate::machine::{Machine, ProgramFactory};
+use crate::compiler::Compiler;
+use crate::machine::Machine;
 
 #[test]
 fn simple_string() {
-    let hir = Parser::new().parse(r"abc").unwrap();
-    let program = visit(&hir, ProgramFactory::default()).unwrap();
+    let program = Compiler::compile(r"abc");
     let mut machine = Machine::new(program);
     assert!(machine.run("abc".to_string()));
     assert!(machine.run("123abc".to_string()));
@@ -15,264 +13,231 @@ fn simple_string() {
 
 #[test]
 fn simple_string_end_matching_should_succeed() {
-    let hir = Parser::new().parse(r"abc$").unwrap();
-    let program = visit(&hir, ProgramFactory::default()).unwrap();
+    let program = Compiler::compile(r"abc");
     let mut machine = Machine::new(program);
     assert!(machine.run("123abc".to_string()));
 }
 
 #[test]
 fn simple_string_end_matching_should_fail() {
-    let hir = Parser::new().parse(r"abc$").unwrap();
-    let program = visit(&hir, ProgramFactory::default()).unwrap();
+    let program = Compiler::compile(r"abc$");
     let mut machine = Machine::new(program);
     assert!(!machine.run("123abc456".to_string()));
 }
 
 #[test]
 fn simple_string_start_matching_should_succeed() {
-    let hir = Parser::new().parse(r"^abc").unwrap();
-    let program = visit(&hir, ProgramFactory::default()).unwrap();
+    let program = Compiler::compile(r"^abc");
     let mut machine = Machine::new(program);
     assert!(machine.run("abc123".to_string()));
 }
 
 #[test]
 fn simple_string_start_matching_should_fail() {
-    let hir = Parser::new().parse(r"^abc").unwrap();
-    let program = visit(&hir, ProgramFactory::default()).unwrap();
+    let program = Compiler::compile(r"^abc");
     let mut machine = Machine::new(program);
     assert!(!machine.run("123abc".to_string()));
 }
 
 #[test]
 fn simple_string_exact_matching_should_succeed() {
-    let hir = Parser::new().parse(r"^abc$").unwrap();
-    let program = visit(&hir, ProgramFactory::default()).unwrap();
+    let program = Compiler::compile(r"^abc$");
     let mut machine = Machine::new(program);
     assert!(machine.run("abc".to_string()));
 }
 
 #[test]
 fn simple_string_exact_matching_should_fail() {
-    let hir = Parser::new().parse(r"^abc$").unwrap();
-    let program = visit(&hir, ProgramFactory::default()).unwrap();
+    let program = Compiler::compile(r"^abc$");
     let mut machine = Machine::new(program);
     assert!(!machine.run("aabc".to_string()));
 }
 
 #[test]
 fn simple_string_exact_matching_should_fail_2() {
-    let hir = Parser::new().parse(r"^abc$").unwrap();
-    let program = visit(&hir, ProgramFactory::default()).unwrap();
+    let program = Compiler::compile(r"^abc$");
     let mut machine = Machine::new(program);
     assert!(!machine.run("abccc".to_string()));
 }
 
 #[test]
 fn simple_string_one_or_more_matching_should_succeed() {
-    let hir = Parser::new().parse(r"^ab+c$").unwrap();
-    let program = visit(&hir, ProgramFactory::default()).unwrap();
+    let program = Compiler::compile(r"^ab+c$");
     let mut machine = Machine::new(program);
     assert!(machine.run("abbc".to_string()));
 }
 
 #[test]
 fn simple_string_one_or_more_matching_should_succeed_2() {
-    let hir = Parser::new().parse(r"^ab+c$").unwrap();
-    let program = visit(&hir, ProgramFactory::default()).unwrap();
+    let program = Compiler::compile(r"^ab+c$");
     let mut machine = Machine::new(program);
     assert!(machine.run("abc".to_string()));
 }
 
 #[test]
 fn simple_string_one_or_more_matching_should_fail() {
-    let hir = Parser::new().parse(r"^ab+c$").unwrap();
-    let program = visit(&hir, ProgramFactory::default()).unwrap();
+    let program = Compiler::compile(r"^ab+c$");
     let mut machine = Machine::new(program);
     assert!(!machine.run("ac".to_string()));
 }
 
 #[test]
 fn simple_string_zero_or_more_matching_should_succeed() {
-    let hir = Parser::new().parse(r"^ab*c$").unwrap();
-    let program = visit(&hir, ProgramFactory::default()).unwrap();
+    let program = Compiler::compile(r"^ab*c$");
     let mut machine = Machine::new(program);
     assert!(machine.run("ac".to_string()));
 }
 
 #[test]
 fn simple_string_zero_or_more_matching_should_succeed_2() {
-    let hir = Parser::new().parse(r"^ab*c$").unwrap();
-    let program = visit(&hir, ProgramFactory::default()).unwrap();
+    let program = Compiler::compile(r"^ab*c$");
     let mut machine = Machine::new(program);
     assert!(machine.run("abbbc".to_string()));
 }
 
 #[test]
 fn simple_string_optional_matching_should_succeed() {
-    let hir = Parser::new().parse(r"^ab?c$").unwrap();
-    let program = visit(&hir, ProgramFactory::default()).unwrap();
+    let program = Compiler::compile(r"^ab?c$");
     let mut machine = Machine::new(program);
     assert!(machine.run("abc".to_string()));
 }
 
 #[test]
 fn simple_string_optional_matching_should_succeed_2() {
-    let hir = Parser::new().parse(r"^ab?c$").unwrap();
-    let program = visit(&hir, ProgramFactory::default()).unwrap();
+    let program = Compiler::compile(r"^ab?c$");
     let mut machine = Machine::new(program);
     assert!(machine.run("ac".to_string()));
 }
 
 #[test]
 fn simple_string_optional_matching_should_fail() {
-    let hir = Parser::new().parse(r"^ab?c$").unwrap();
-    let program = visit(&hir, ProgramFactory::default()).unwrap();
+    let program = Compiler::compile(r"^ab?c$");
     let mut machine = Machine::new(program);
     assert!(!machine.run("abbc".to_string()));
 }
 
 #[test]
 fn simple_string_numbered_matching_should_succeed() {
-    let hir = Parser::new().parse(r"^ab{2}c$").unwrap();
-    let program = visit(&hir, ProgramFactory::default()).unwrap();
+    let program = Compiler::compile(r"^ab{2}c$");
     let mut machine = Machine::new(program);
     assert!(machine.run("abbc".to_string()));
 }
 
 #[test]
 fn simple_string_numbered_matching_should_fail() {
-    let hir = Parser::new().parse(r"^ab{2}c$").unwrap();
-    let program = visit(&hir, ProgramFactory::default()).unwrap();
+    let program = Compiler::compile(r"^ab{2}c$");
     let mut machine = Machine::new(program);
     assert!(!machine.run("abbbc".to_string()));
 }
 
 #[test]
 fn simple_string_numbered_matching_should_fail_2() {
-    let hir = Parser::new().parse(r"^ab{2}c$").unwrap();
-    let program = visit(&hir, ProgramFactory::default()).unwrap();
+    let program = Compiler::compile(r"^ab{2}c$");
     let mut machine = Machine::new(program);
     assert!(!machine.run("abc".to_string()));
 }
 
 #[test]
 fn simple_string_numbered_matching_should_succeed_2() {
-    let hir = Parser::new().parse(r"^ab{3,}c$").unwrap();
-    let program = visit(&hir, ProgramFactory::default()).unwrap();
+    let program = Compiler::compile(r"^ab{3,}c$");
     let mut machine = Machine::new(program);
     assert!(machine.run("abbbc".to_string()));
 }
 
 #[test]
 fn simple_string_numbered_matching_should_succeed_3() {
-    let hir = Parser::new().parse(r"^ab{3,}c$").unwrap();
-    let program = visit(&hir, ProgramFactory::default()).unwrap();
+    let program = Compiler::compile(r"^ab{3,}c$");
     let mut machine = Machine::new(program);
     assert!(machine.run("abbbbbbc".to_string()));
 }
 
 #[test]
 fn simple_string_numbered_matching_should_fail_3() {
-    let hir = Parser::new().parse(r"^ab{3,}c$").unwrap();
-    let program = visit(&hir, ProgramFactory::default()).unwrap();
+    let program = Compiler::compile(r"^ab{3,}c$");
     let mut machine = Machine::new(program);
     assert!(!machine.run("abbc".to_string()));
 }
 
 #[test]
 fn simple_string_numbered_matching_should_succeed_4() {
-    let hir = Parser::new().parse(r"^ab{2,4}c$").unwrap();
-    let program = visit(&hir, ProgramFactory::default()).unwrap();
+    let program = Compiler::compile(r"^ab{2,4}c$");
     let mut machine = Machine::new(program);
     assert!(machine.run("abbbbc".to_string()));
 }
 
 #[test]
 fn simple_string_numbered_matching_should_fail_4() {
-    let hir = Parser::new().parse(r"^ab{2,4}c$").unwrap();
-    let program = visit(&hir, ProgramFactory::default()).unwrap();
+    let program = Compiler::compile(r"^ab{2,4}c$");
     let mut machine = Machine::new(program);
     assert!(!machine.run("abc".to_string()));
 }
 
 #[test]
 fn simple_string_numbered_matching_should_fail_5() {
-    let hir = Parser::new().parse(r"^ab{2,4}c$").unwrap();
-    let program = visit(&hir, ProgramFactory::default()).unwrap();
+    let program = Compiler::compile(r"^ab{2,4}c$");
     let mut machine = Machine::new(program);
     assert!(!machine.run("abbbbbc".to_string()));
 }
 
 #[test]
 fn escaping_special_characters_should_succeed() {
-    let hir = Parser::new().parse(r"^\.$").unwrap();
-    let program = visit(&hir, ProgramFactory::default()).unwrap();
+    let program = Compiler::compile(r"^\.$");
     let mut machine = Machine::new(program);
     assert!(machine.run(".".to_string()));
 }
 
 #[test]
 fn escaping_special_characters_should_succeed_2() {
-    let hir = Parser::new().parse(r"^\*$").unwrap();
-    let program = visit(&hir, ProgramFactory::default()).unwrap();
+    let program = Compiler::compile(r"^\*$");
     let mut machine = Machine::new(program);
     assert!(machine.run("*".to_string()));
 }
 
 #[test]
 fn character_range_matching_should_succeed() {
-    let hir = Parser::new().parse(r"^[abc]$").unwrap();
-    let program = visit(&hir, ProgramFactory::default()).unwrap();
+    let program = Compiler::compile(r"^[abc]$");
     let mut machine = Machine::new(program);
     assert!(machine.run("a".to_string()));
 }
 
 #[test]
 fn character_range_matching_should_fail() {
-    let hir = Parser::new().parse(r"^[abc]$").unwrap();
-    let program = visit(&hir, ProgramFactory::default()).unwrap();
+    let program = Compiler::compile(r"^[abc]$");
     let mut machine = Machine::new(program);
     assert!(!machine.run("d".to_string()));
 }
 
 #[test]
 fn character_range_not_matching_should_succeed() {
-    let hir = Parser::new().parse(r"^[^ade]$").unwrap();
-    let program = visit(&hir, ProgramFactory::default()).unwrap();
+    let program = Compiler::compile(r"^[^ade]$");
     let mut machine = Machine::new(program);
     assert!(machine.run("b".to_string()));
 }
 
 #[test]
 fn character_range_not_matching_should_fail() {
-    let hir = Parser::new().parse(r"^[^ade]$").unwrap();
-    let program = visit(&hir, ProgramFactory::default()).unwrap();
+    let program = Compiler::compile(r"^[^ade]$");
     let mut machine = Machine::new(program);
     assert!(!machine.run("a".to_string()));
 }
 
 #[test]
 fn any_character_matching_should_succeed() {
-    let hir = Parser::new().parse(r"^.$").unwrap();
-    let program = visit(&hir, ProgramFactory::default()).unwrap();
+    let program = Compiler::compile(r"^.$");
     let mut machine = Machine::new(program);
     assert!(machine.run("A".to_string()));
 }
 
 #[test]
 fn case_insensitive_argument_should_succeed() {
-    let hir = Parser::new().parse(r"(?i)^abc$").unwrap();
-    let program = visit(&hir, ProgramFactory::default()).unwrap();
+    let program = Compiler::compile(r"(?i)^abc$");
     let mut machine = Machine::new(program);
     assert!(machine.run("ABC".to_string()));
 }
 
 #[test]
 fn alternation_should_succeed() {
-    let hir = Parser::new().parse(r"0a|bcd$").unwrap();
-    let program = visit(&hir, ProgramFactory::default()).unwrap();
+    let program = Compiler::compile(r"0a|bcd$");
     let mut machine = Machine::new(program);
     assert!(machine.run("0a".to_string()));
     machine.reset();
@@ -281,8 +246,7 @@ fn alternation_should_succeed() {
 
 #[test]
 fn alternation_should_succeed_2() {
-    let hir = Parser::new().parse(r"a(bc|ed)42$").unwrap();
-    let program = visit(&hir, ProgramFactory::default()).unwrap();
+    let program = Compiler::compile(r"a(bc|ed)42$");
     let mut machine = Machine::new(program);
     assert!(machine.run("abc42".to_string()));
     machine.reset();
@@ -292,8 +256,7 @@ fn alternation_should_succeed_2() {
 
 #[test]
 fn alternation_should_fail() {
-    let hir = Parser::new().parse(r"0a|bcd$").unwrap();
-    let program = visit(&hir, ProgramFactory::default()).unwrap();
+    let program = Compiler::compile(r"0a|bcd$");
     let mut machine = Machine::new(program);
     assert!(!machine.run("0b".to_string()));
     machine.reset();
@@ -302,8 +265,7 @@ fn alternation_should_fail() {
 
 #[test]
 fn alternation_should_fail_2() {
-    let hir = Parser::new().parse(r"a(bc|ed)42$").unwrap();
-    let program = visit(&hir, ProgramFactory::default()).unwrap();
+    let program = Compiler::compile(r"a(bc|ed)42$");
     let mut machine = Machine::new(program);
     assert!(!machine.run("abd42".to_string()));
     machine.reset();
@@ -312,8 +274,7 @@ fn alternation_should_fail_2() {
 
 #[test]
 fn alternation_string_numbered_matching_should_succeed() {
-    let hir = Parser::new().parse(r"^hel(ab{2}|l{3,}o)bc$").unwrap();
-    let program = visit(&hir, ProgramFactory::default()).unwrap();
+    let program = Compiler::compile(r"^hel(ab{2}|l{3,}o)bc$");
     let mut machine = Machine::new(program);
     assert!(machine.run("helabbbc".to_string()));
     machine.reset();
@@ -322,8 +283,7 @@ fn alternation_string_numbered_matching_should_succeed() {
 
 #[test]
 fn alternation_string_numbered_matching_should_fail() {
-    let hir = Parser::new().parse(r"^hel(ab{2}|l{3,}o)bc$").unwrap();
-    let program = visit(&hir, ProgramFactory::default()).unwrap();
+    let program = Compiler::compile(r"^hel(ab{2}|l{3,}o)bc$");
     let mut machine = Machine::new(program);
     assert!(!machine.run("helabbc".to_string()));
     machine.reset();
